Consolidate utils into Constants.kt, migrate tests
Merge AspectRatio.kt and SongHeaderLines.kt into DisplayGeometry.kt (renamed Constants.kt). Move ConstantsTest and UpdateCheckIntervalTest from composeApp to :settings, and move ClockFormatTest's announcements test to its correct home. Update detekt baseline and settings/AGENT.md to reflect the new file layout.
