+++ composeApp/src/jvmMain/kotlin/org/churchpresenter/app/churchpresenter/utils/Constants.kt
@@ -12,8 +12,10 @@ import java.awt.GraphicsDevice
 import java.awt.GraphicsEnvironment
 import java.awt.HeadlessException
 import java.awt.Rectangle
+import java.util.Locale
 
 private const val SCREEN_POLL_INTERVAL_MS = 2000L
+private const val MAX_SIMPLE_RATIO_SIDE = 64
 
 /** Empty on a headless JVM (CI, or a genuinely displayless deployment) instead of throwing. */
 private fun safeScreenDevices(): Array<GraphicsDevice> = try {
@@ -105,3 +107,34 @@ fun findScreenIndexByBounds(screens: Array<GraphicsDevice>, x: Int, y: Int, w: I
         b.x == x && b.y == y && b.width == w && b.height == h
     }.takeIf { it >= 0 }
 }
+
+/** Returns the aspect ratio of the presenter screen. */
+fun presenterAspectRatio(): Float = aspectRatioOf(presenterScreenBounds())
+
+internal fun aspectRatioOf(bounds: Rectangle): Float =
+    bounds.width.toFloat() / bounds.height.toFloat()
+
+/** Formats an aspect ratio as a common name (e.g. "16:9") or decimal fallback (e.g. "1.78:1"). */
+fun formatAspectRatio(width: Int, height: Int): String {
+    val gcd = gcd(width, height)
+    val w = width / gcd
+    val h = height / gcd
+    // Accept simplified ratios where both sides are reasonable (≤64)
+    return if (w <= MAX_SIMPLE_RATIO_SIDE && h <= MAX_SIMPLE_RATIO_SIDE) "$w:$h"
+    else String.format(Locale.US, "%.2f:1", width.toFloat() / height.toFloat())
+}
+
+private fun gcd(a: Int, b: Int): Int = if (b == 0) a else gcd(b, a % b)
+
+/**
+ * Any line wrapped in [] or {} is a section header — except one holding nothing but a chord, which
+ * an instrumental break writes as its own line. See [ChordTransposer.isSectionHeader].
+ */
+fun isHeaderLine(line: String): Boolean = ChordTransposer.isSectionHeader(line)
+
+/** {} = chorus, [] = verse/other */
+fun isChorusHeader(line: String): Boolean {
+    val t = line.trim()
+    return t.startsWith("{") && t.endsWith("}")
+}
+
